requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
sqlalchemy==1.4.41
pymysql==1.0.3
python-dotenv==1.0.0
passlib==1.7.4
python-jose==3.3.0
pydantic==1.10.4